Update task, target tests
