+++ config/standard/test/targets.spec.js
@@ -1289,6 +1289,279 @@ describe('Standard Configuration Targets', function() {
     });
   });
 
+  describe('Nutrition screening by CHW', function() {
+    it('should create a child screened target instance', function() {
+      // given
+      session.assert(childWithReport(
+        {
+          form: 'G',
+          fields: {},
+          reported_date: today,
+        }
+      ));
+
+      // expect
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-nutrition-children-screened-growth-monitoring',
+            deleted: false,
+            type: 'nutrition-children-screened-growth-monitoring',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+        });
+    });
+
+  });
+
+  describe('Nutrition screening at facility', function(){
+
+    it('should create underweight target instance', function() {
+
+      const r = {
+        form: 'nutrition_screening',
+        fields: {
+          zscore: {
+            zscore_wfa: -3
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      // expect
+      return session.emitTargets()
+        .then(targets => {
+          const expectedTarget = {
+            _id: 'child-1-nutrition-children-underweight',
+            deleted: false,
+            type: 'nutrition-children-underweight',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+
+        });
+    });
+
+
+    it('should create stunted growth target instance', function() {
+      const r = {
+        form: 'nutrition_screening',
+        fields: {
+          zscore: {
+            zscore_hfa: -3
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-children-stunted',
+            deleted: false,
+            type: 'children-stunted',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+
+        });
+    });
+
+  });
+
+  describe('children active MAM', function(){
+    it('should create active MAM target for WFH z-score between -3 & -2', function(){
+      const r = {
+        form: 'treatment_enrollment',
+        fields: {
+          zscore: {
+            zscore_wfh: -3
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-children-mam',
+            deleted: false,
+            type: 'children-mam',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+        });
+
+    });
+
+    it('should create active MAM trget for MUAC between 11.5 & 12.4 cm', function(){
+      const r = {
+        form: 'treatment_enrollment',
+        fields: {
+          zscore: {
+            muac: 12
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-children-mam',
+            deleted: false,
+            type: 'children-mam',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+        });
+    });
+  });
+
+  describe('children active SAM', function(){
+    it('should create active SAM target for WFH z-score less than -3', function(){
+      const r = {
+        form: 'treatment_enrollment',
+        fields: {
+          zscore: {
+            zscore_wfh: -4
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-children-sam',
+            deleted: false,
+            type: 'children-sam',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+        });
+
+    });
+
+    it('should create active SAM trget for MUAC less than 11.5 cm', function(){
+      const r = {
+        form: 'treatment_enrollment',
+        fields: {
+          zscore: {
+            muac: 10
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-children-sam',
+            deleted: false,
+            type: 'children-sam',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+        });
+    });
+  });
+
+  describe('children active OTP', function(){
+    it('should create active OTP for children enrolled', function(){
+      const r = {
+        form: 'treatment_enrollment',
+        fields: {
+          enrollment: {
+            program: 'OTP'
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-children-otp',
+            deleted: false,
+            type: 'children-otp',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+        });
+
+    });
+  });
+
+  describe('children active SFP', function(){
+    it('should create active SFP target for children enrolled', function(){
+      const r = {
+        form: 'treatment_enrollment',
+        fields: {
+          enrollment: {
+            program: 'SFP'
+          }
+        },
+        reported_date: today,
+      };
+
+      session.assert(childWithReport(r));
+
+      return session.emitTargets()
+        .then(targets => {
+
+          const expectedTarget = {
+            _id: 'child-1-children-sfp',
+            deleted: false,
+            type: 'children-sfp',
+            pass: true,
+            date: today
+          };
+
+          assert.deepInclude(targets, expectedTarget);
+        });
+
+    });
+  });
+
   function childWithNoReports() {
     return childWithReports();
   }
